--- src/components/overlays/overlay-section.styl
@@ -1,11 +1,0 @@
-@import '../constants.styl'
-
-.section
-  margin: 0
-  padding: 14px 12px
-  color: primary
-  &:not(:last-child)
-    border-bottom: 1px solid secondary
-
-.info
-  background-color: secondary-background
