--- src/components/overlays/overlay-title.styl
@@ -1,7 +1,0 @@
-@import '../constants.styl'
-
-.title
-  font-size: inherit
-  font-weight: bold
-  margin: 0
-
