+++ src/components/overlays/overlays.styl
@@ -23,3 +23,17 @@ pop-shadow = 0px 2px 5px 0px rgba(0,0,0,0.27), 0px 1px 1px 0px rgba(0,0,0,0.15)
   box-shadow: pop-shadow
   z-index: 15
   overflow: auto
+
+.section
+  margin: 0
+  padding: 14px 12px
+  color: primary
+  &:not(:last-child)
+    border-bottom: 1px solid secondary
+ &.info
+  background-color: secondary-background
+
+.title
+  font-size: inherit
+  font-weight: bold
+  margin: 0
